+++ veltahleon/src/main/java/com/testament/veltahleon/model/entities/history/Event.java
@@ -1,0 +1,33 @@
+package com.testament.veltahleon.model.entities.history;
+
+import com.testament.veltahleon.model.entities.calendar.Year;
+import jakarta.persistence.*;
+import lombok.*;
+
+@Entity
+@Table(name = "events")
+@Getter
+@Setter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@EqualsAndHashCode
+@ToString
+public class Event {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String incident;
+
+    @ManyToOne(fetch = FetchType.LAZY, cascade = {CascadeType.REFRESH,
+            CascadeType.DETACH,
+            CascadeType.MERGE,
+            CascadeType.PERSIST})
+    @JoinColumn(name = "year_id")
+    private Year year;
+
+    @Column(name = "description", columnDefinition = "text")
+    private String description;
+}