+++ veltahleon/src/main/java/com/testament/veltahleon/services/ifc/society/PersonService.java
@@ -14,5 +14,5 @@ public interface PersonService {
     Boolean deletePersonByID(Long id);
     Person savePerson(Person person);
     Person updatePerson(Long id, Person person);
-    Person updatePerson(Person person);
+    Person modifyPerson(Long id, Person person);
 }